+++ src/api/instance.service.ts
@@ -1,4 +1,5 @@
 export * from '@/api/modules/common'
+export * from '@/api/modules/account'
 export * from '@/api/modules/deploy'
 export * from '@/api/modules/crm'
 export * from '@/api/modules/finance'