+++ src/api/modules/crm/modules/client.service.ts
@@ -1,23 +1,5 @@
 import { request } from '@/utils'
 
-/**销售管理-客户列表-新增客户**/
-export function httpBaseCrmClientCommonCreate(data: Omix) {
-    return request({
-        url: `/api/consumers/create`,
-        method: 'POST',
-        data
-    })
-}
-
-/**销售管理-客户列表-分页列表查询**/
-export function httpBaseCrmClientCommonConsumer(data: Omix) {
-    return request({
-        url: `/api/consumers/column`,
-        method: 'POST',
-        data
-    })
-}
-
 /**销售管理-客户列表-客户详情**/
 export function httpBaseCrmClientResolver(data: Omix) {
     return request({